overflow

+++ crates/brontes-types/src/db/cex/trades/time_window_vwam.rs
@@ -461,7 +461,7 @@ impl<'a> TimeWindowTrades<'a> {
                 .max_by_key(|t| t.timestamp)
                 .map(|t| t.timestamp);
 
-            let time_before = min.clone().map(|min| block_timestamp - min);
+            let time_before = min.clone().map(|min| block_timestamp as isize - min as isize);
             let time_after = max
                 .clone()
                 .map(|max| max as isize - block_timestamp as isize);
